@@ -2,6 +2,7 @@
 
 namespace Prescreen\ApiResourceBundle\Tests\Application\Services;
 
+use PHPUnit\Framework\MockObject\MockObject;
 use Prescreen\ApiResourceBundle\Application\ApiResources\ExampleResource;
 use Prescreen\ApiResourceBundle\Application\Configuration\FieldOptions\BoolField;
 use Prescreen\ApiResourceBundle\Application\Services\ApiResourceTransformer;
@@ -10,14 +11,8 @@ use PHPUnit\Framework\TestCase;
 
 class ApiResourceTransformerRegistryTest extends TestCase
 {
-    /**
-     * @var iterable
-     */
-    private $apiResourceTransformers;
-    /**
-     * @var ApiResourceTransformerRegistry
-     */
-    private $testService;
+    private iterable $apiResourceTransformers;
+    private ApiResourceTransformerRegistry $testService;
 
     public function setUp(): void
     {
@@ -27,30 +22,19 @@ class ApiResourceTransformerRegistryTest extends TestCase
         $this->testService = new ApiResourceTransformerRegistry($this->apiResourceTransformers);
     }
 
-    /**
-     * @test
-     */
-    public function givenResourceTransformerForGivenResourceClassHasNotBeenRegisteredThenThrowException()
+    public function testItThrowsExceptionIfResourceTransformerForGivenResourceClassHasNotBeenRegistered(): void
     {
         $this->expectException(\InvalidArgumentException::class);
 
         $this->testService->get(BoolField::class);
     }
 
-    /**
-     * @test
-     */
-    public function givenResourceClassThenReturnResourceTransformer()
+    public function testItReturnsResourceTransformerForGivenResourceClass(): void
     {
         $this->assertInstanceOf(ApiResourceTransformer::class, $this->testService->get(ExampleResource::class));
     }
 
-    /**
-     * @param string $resourceClass
-     *
-     * @return ApiResourceTransformer
-     */
-    private function createInternalApiResourceTransformer(string $resourceClass): ApiResourceTransformer
+    private function createInternalApiResourceTransformer(string $resourceClass): ApiResourceTransformer&MockObject
     {
         $resourceTransformer = $this->createMock(ApiResourceTransformer::class);
         $resourceTransformer->method('getResourceClass')->willReturn($resourceClass);
