@@ -17,18 +17,9 @@ use PHPUnit\Framework\TestCase;
 
 class IdCollectionValidatorTest extends TestCase
 {
-    /**
-     * @var EntityManagerInterface
-     */
-    protected $em;
-    /**
-     * @var EntityRepository
-     */
-    protected $repository;
-    /**
-     * @var IdCollectionValidator
-     */
-    private $testService;
+    protected EntityManagerInterface $em;
+    protected EntityRepository $repository;
+    private IdCollectionValidator $testService;
 
     public function setUp(): void
     {
@@ -36,70 +27,40 @@ class IdCollectionValidatorTest extends TestCase
         $this->testService = new IdCollectionValidator($this->em);
     }
 
-    /**
-     * @test
-     */
-    public function givenFieldOptionsAreNotOfTypeIdFieldThenThrowException()
+    public function testItThrowsExceptionIfGivenFieldOptionsAreNotOfTypeIdField(): void
     {
         $this->expectException(FieldTypeException::class);
 
         $this->testService->validate('example_ids', [1], new IntField());
     }
 
-    /**
-     * @test
-     */
-    public function givenValueIsNotOfTypeArrayThenThrowException()
+    public function testItThrowsExceptionIfValueIsNotOfTypeArray(): void
     {
         $this->expectException(FieldTypeException::class);
 
         $this->testService->validate('example_ids', 'foo', new IdCollectionField(ExampleEntity::class));
     }
 
-    /**
-     * @test
-     */
-    public function givenEntityCannotBeFoundThenThrowException()
+    public function testItThrowsExceptionIfEntityCannotBeFound(): void
     {
         $this->expectException(LinkedObjectNotFoundException::class);
 
         $repository = $this->expectRepositoryToBeFound();
-        $repository->expects($this->once())->method('find')->willReturn(null);
+        $repository->expects($this->once())->method('findOneBy')->willReturn(null);
 
         $this->testService->validate('example_ids', [23502357], new IdCollectionField(ExampleEntity::class));
     }
 
-    /**
-     * @test
-     */
-    public function givenFieldOptionsHaveRecruiterAccountThenFindEntitiesWithRecruiterAccountAndReturnThem()
-    {
-        $repository = $this->expectRepositoryToBeFound();
-        $exampleEntity = new ExampleEntity();
-        $repository->expects($this->exactly(2))->method('find')->willReturn($exampleEntity);
-
-        $this->assertEquals(new ArrayCollection([$exampleEntity, $exampleEntity]), $this->testService->validate('example_ids', [1, 2], new IdCollectionField(ExampleEntity::class, true)));
-    }
-
-    /**
-     * @test
-     */
-    public function givenFieldIsNotRequiredAndValueIsNullThenReturnNull()
+    public function testItReturnsNullIfFieldIsNotRequiredAndValueIsNull(): void
     {
         $this->assertNull($this->testService->validate('example_ids', null, new IdCollectionField(ExampleEntity::class, false)));
     }
 
-    /**
-     * @test
-     */
-    public function itIsOfTypeIdCollection()
+    public function testItIsOfTypeIdCollection(): void
     {
         $this->assertSame(IdCollectionField::TYPE, $this->testService->getType());
     }
 
-    /**
-     * @return MockObject
-     */
     private function expectRepositoryToBeFound(): MockObject
     {
         $repository = $this->createMock(ExampleRepository::class);
