@@ -15,14 +15,8 @@ use PHPUnit\Framework\TestCase;
 
 class IdValidatorTest extends TestCase
 {
-    /**
-     * @var EntityManagerInterface
-     */
-    protected $em;
-    /**
-     * @var IdValidator
-     */
-    private $testService;
+    protected EntityManagerInterface $em;
+    private IdValidator $testService;
 
     public function setUp(): void
     {
@@ -30,30 +24,23 @@ class IdValidatorTest extends TestCase
         $this->testService = new IdValidator($this->em);
     }
 
-    /**
-     * @test
-     */
-    public function givenFieldOptionsAreNotOfTypeIdFieldThenThrowException()
+    public function testItThrowsExceptionIfGivenFieldOptionsAreNotOfTypeIdField(): void
     {
         $this->expectException(FieldTypeException::class);
 
         $this->testService->validate('department_id', 1, new IntField());
     }
 
-    /**
-     * @test
-     */
-    public function givenValueIsNotOfTypeIntThenThrowException()
+    public function testItThrowsExceptionIfValueIsNotOfTypeInt(): void
     {
         $this->expectException(FieldTypeException::class);
+
+        $this->expectRepositoryToBeFound();
 
         $this->testService->validate('example_id', 'foo', new IdField(ExampleEntity::class));
     }
 
-    /**
-     * @test
-     */
-    public function givenEntityCannotBeFoundThenThrowException()
+    public function testItThrowsExceptionIfEntityCannotBeFound(): void
     {
         $this->expectException(LinkedObjectNotFoundException::class);
 
@@ -63,42 +50,14 @@ class IdValidatorTest extends TestCase
         $this->testService->validate('example_id', 23502357, new IdField(ExampleEntity::class));
     }
 
-    /**
-     * @test
-     */
-    public function givenFieldOptionsHaveRecruiterAccountThenFindEntityWithRecruiterAccountAndReturnIt()
+    public function testItReturnsNullIfFieldIsNotRequiredAndValueIsNull(): void
     {
-        $repository = $this->expectRepositoryToBeFound();
-        $exampleEntity = new ExampleEntity();
-        $repository->expects($this->once())->method('findOneBy')->willReturn($exampleEntity);
+        $this->expectRepositoryToBeFound();
 
-        $this->assertEquals($exampleEntity, $this->testService->validate('example_id', 23502357, new IdField(ExampleEntity::class, true)));
-    }
-
-    /**
-     * @test
-     */
-    public function givenFieldOptionsDoNotHaveRecruiterAccountThenFindEntityAndReturnIt()
-    {
-        $repository = $this->expectRepositoryToBeFound();
-        $department = new ExampleEntity();
-        $repository->expects($this->once())->method('findOneBy')->willReturn($department);
-
-        $this->assertEquals($department, $this->testService->validate('example_id', 23502357, new IdField(ExampleEntity::class)));
-    }
-
-    /**
-     * @test
-     */
-    public function givenFieldIsNotRequiredAndValueIsNullThenReturnNull()
-    {
         $this->assertNull($this->testService->validate('example_id', null, new IdField(false)));
     }
 
-    /**
-     * @test
-     */
-    public function itIsOfTypeId()
+    public function testItIsOfTypeId(): void
     {
         $this->assertSame(IdField::TYPE, $this->testService->getType());
     }
