@@ -9,54 +9,36 @@ use Prescreen\ApiResourceBundle\Exception\FieldTypeException;
 
 class JsonValidatorTest extends TestCase
 {
-    /**
-     * @var JsonValidator
-     */
-    private $testService;
+    private JsonValidator $testService;
 
     protected function setUp(): void
     {
         $this->testService = new JsonValidator();
     }
 
-    /**
-     * @test
-     */
-    public function givenValueIsNotOfTypeJsonThenThrowException()
+    public function testItThrowsExceptionIfValueIsNotOfTypeJson(): void
     {
         $this->expectException(FieldTypeException::class);
 
         $this->testService->validate('json', false, new JsonField());
     }
 
-    /**
-     * @test
-     */
-    public function givenValidJsonValueThenReturnIt()
+    public function testItReturnsValueIfJsonIsValid(): void
     {
         $this->assertSame(json_encode(['header' => 'Welcome']), $this->testService->validate('json', json_encode(['header' => 'Welcome']), new JsonField()));
     }
 
-    /**
-     * @test
-     */
-    public function givenValidArrayValueThenReturnIt()
+    public function testItReturnsValueIfGivenAValidArray(): void
     {
         $this->assertSame(['header' => 'Welcome'], $this->testService->validate('json', ['header' => 'Welcome'], new JsonField()));
     }
 
-    /**
-     * @test
-     */
-    public function givenFieldIsNotRequiredAndValueIsNullThenReturnNull()
+    public function testItReturnsNullIfFieldIsNotRequiredAndValueIsNull(): void
     {
         $this->assertNull($this->testService->validate('json', null, new JsonField()));
     }
 
-    /**
-     * @test
-     */
-    public function itIsOfTypeString()
+    public function testItIsOfTypeString(): void
     {
         $this->assertSame(JsonField::TYPE, $this->testService->getType());
     }
